move single file folders into the components directory
