+++ packages/editor/src/components/Toolbar.tsx
@@ -40,16 +40,16 @@ import { ChevronDownSm, File04Sm, SquaresLg, UploadCloud02Sm } from '@ir-engine/
 import { t } from 'i18next'
 import React, { Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
-import { confirmSceneExists, onNewScene, onSaveScene, saveSceneGLTF } from '../../functions/sceneFunctions'
-import { cmdOrCtrlString } from '../../functions/utils'
-import { uploadFiles } from '../../panels/assets/topbar'
-import { EditorState } from '../../services/EditorServices'
-import { UIAddonsState } from '../../services/UIAddonsState'
-import CreatePrefabPanel from '../dialogs/CreatePrefabPanelDialog'
-import CreateSceneDialog from '../dialogs/CreateScenePanelDialog'
-import ImportSettingsPanel from '../dialogs/ImportSettingsPanelDialog'
-import SaveNewSceneDialog from '../dialogs/SaveNewSceneDialog'
-import QuitToDashboardConfirmationDialog from './../dialogs/QuitToDashboardConfirmationDialog'
+import { confirmSceneExists, onNewScene, onSaveScene, saveSceneGLTF } from '../functions/sceneFunctions'
+import { cmdOrCtrlString } from '../functions/utils'
+import { uploadFiles } from '../panels/assets/topbar'
+import { EditorState } from '../services/EditorServices'
+import { UIAddonsState } from '../services/UIAddonsState'
+import CreatePrefabPanel from './dialogs/CreatePrefabPanelDialog'
+import CreateSceneDialog from './dialogs/CreateScenePanelDialog'
+import ImportSettingsPanel from './dialogs/ImportSettingsPanelDialog'
+import QuitToDashboardConfirmationDialog from './dialogs/QuitToDashboardConfirmationDialog'
+import SaveNewSceneDialog from './dialogs/SaveNewSceneDialog'
 
 const onImportAsset = async () => {
   const { projectName } = getState(EditorState)
