+++ packages/editor/src/hooks/useEditorActiveInstances.ts
