+++ packages/editor/src/panels/hierarchy/hierarchynode.tsx
@@ -64,7 +64,7 @@ import { MdKeyboardArrowDown, MdKeyboardArrowRight } from 'react-icons/md'
 import { PiEyeBold, PiEyeClosedBold, PiLockBold, PiLockOpenBold } from 'react-icons/pi'
 import { ListChildComponentProps } from 'react-window'
 import { twMerge } from 'tailwind-merge'
-import { IconComponent } from '../../components/panels/IconComponent'
+import { IconComponent } from '../../components/IconComponent'
 import { exportRelativeGLTF } from '../../functions/exportGLTF'
 import { EditorHelperState, PlacementMode } from '../../services/EditorHelperState'
 import { EditorHistoryFunctions } from '../../services/EditorHistoryState'
